+++ CHANGELOG.md
@@ -8,15 +8,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unpublished]
 
 ### Added
+
 - Scripts: localnet and faucet process start/stop/status
 - Scripts: localnet/devnet/testnet, lsui/dsui/tsui shortcuts
 - Scripts: asui for user selectable active workdir (look for 'set-active' option).
 - Scripts: csui for "cargobin" workdir created when ./cargo/bin/sui exists.
 - Python: 'sysinfo' and 'coinage' demo
 - Rust: 'demo-app'
+- Python: demo app(s) added: `sysinfo` and `coinage`
 
 ### Fixed
 
 ### Changed
+
+- Python requirements.txt updated to use `pysui 0.15.0`
+- Added sysinfo code for Sui 0.29.1 types
 
 ### Removed